requirements-dev.txt
# Development requirements for olgFeast
# Include base requirements
-r requirements.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code quality
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0
django-stubs==5.1.0

# Development tools
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.29.0
jupyter==1.1.1

# Security
bandit==1.7.10
safety==3.2.3

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.1.1

# Performance
django-silk==5.1.0
django-query-profiler==0.6.0
